+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Changelog
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html). 
 
+## [0.1.1] - 2024-01-15
+### Added
+- translation module
+- new translation constants
+- new pytest tests
+- new chunker that chunks by sentences
+
+### Changed
+- moved some functions around for better organization
+- updated walkthrough.ipynb to reflect new capabilities
+
 ## [0.1.0] - 2024-01-14
 ### Added
 - Initial release of Kleio